+++ mobile/lib/screens/quiz/quiz_widgets.dart
@@ -195,235 +195,6 @@ class QuizActionButton extends StatelessWidget {
   }
 }
 
-class QuizQuestionsWidget extends StatefulWidget {
-  const QuizQuestionsWidget(this.quiz, {super.key});
-  final Quiz quiz;
-  @override
-  State<QuizQuestionsWidget> createState() => _QuizQuestionsWidgetState();
-}
-
-class _QuizQuestionsWidgetState extends State<QuizQuestionsWidget> {
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<CurrentQuizQuestionCubit, QuizQuestion?>(
-      builder: (context, state) {
-        if (state == null) {
-          Navigator.pop(context, true);
-          return Text(
-            AppLocalizations.of(context)!.noQuestions,
-          );
-        }
-        return QuizQuestionWidget(
-          state,
-          parentContent: context,
-          quiz: widget.quiz,
-        );
-      },
-    );
-  }
-}
-
-class QuizQuestionWidget extends StatefulWidget {
-  const QuizQuestionWidget(
-    this.currentQuestion, {
-    super.key,
-    required this.parentContent,
-    required this.quiz,
-  });
-  final BuildContext parentContent;
-  final QuizQuestion currentQuestion;
-  final Quiz quiz;
-
-  @override
-  State<QuizQuestionWidget> createState() => _QuizQuestionWidgetState();
-}
-
-class _QuizQuestionWidgetState extends State<QuizQuestionWidget> {
-  bool showAnswer = false;
-  late QuizQuestionOption selectedOption;
-
-  @override
-  void initState() {
-    super.initState();
-    selectedOption = widget.currentQuestion.options[0];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: MediaQuery.of(context).size.height * 0.93,
-      child: Padding(
-        padding: const EdgeInsets.fromLTRB(2, 2, 2, 0),
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          mainAxisAlignment: MainAxisAlignment.start,
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            const SizedBox(
-              height: 1,
-            ),
-            Padding(
-              padding: const EdgeInsets.all(4.0),
-              child: SizedBox(
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    const AppBackButton(),
-                    const SizedBox(
-                      width: 10,
-                    ),
-                    const Padding(
-                      padding: EdgeInsets.all(4.0),
-                      child: QuizDraggingHandle(),
-                    ),
-                    const SizedBox(
-                      width: 10,
-                    ),
-                    Padding(
-                      padding: const EdgeInsets.all(6.0),
-                      child: InkWell(
-                        onTap: () async {
-                          Navigator.pop(widget.parentContent);
-                        },
-                        child: SvgPicture.asset(
-                          'assets/icon/close.svg',
-                          height: 35,
-                          width: 35,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-            const SizedBox(
-              height: 25,
-              child: AutoSizeText(
-                'Air Quality Quiz',
-                style: TextStyle(
-                  color: Color.fromARGB(255, 0, 0, 0),
-                  fontSize: 20,
-                  fontWeight: FontWeight.w500,
-                  //height: 1.50,
-                ),
-              ),
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            const Padding(
-              padding: EdgeInsets.all(8.0),
-              child: QuizProgressBar(),
-            ),
-            const SizedBox(
-              height: 20,
-            ),
-            Visibility(
-              visible: showAnswer,
-              child: QuizAnswerWidget(selectedOption, quiz: widget.quiz,
-                  nextButtonClickCallback: () {
-                int currentIndex =
-                    widget.quiz.questions.indexOf(widget.currentQuestion);
-                if (currentIndex + 1 == widget.quiz.questions.length) {
-                  context.read<CurrentQuizQuestionCubit>().setQuestion(null);
-                } else {
-                  QuizQuestion nextQuestion =
-                      widget.quiz.questions[currentIndex + 1];
-                  context
-                      .read<CurrentQuizQuestionCubit>()
-                      .setQuestion(nextQuestion);
-                }
-                setState(() => showAnswer = false);
-              }),
-            ),
-            Visibility(
-              visible: !showAnswer,
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  SizedBox(
-                    height: 25,
-                    child: AutoSizeText(
-                      widget.currentQuestion.category,
-                      style: const TextStyle(
-                        color: Color.fromARGB(117, 0, 0, 0),
-                        fontSize: 10,
-                        fontWeight: FontWeight.w500,
-                        //height: 1.50,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(
-                    height: 10,
-                  ),
-                  Container(
-                    padding: const EdgeInsets.fromLTRB(30, 0, 30, 0),
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.center,
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        AutoSizeText(
-                          textAlign: TextAlign.center,
-                          maxLines: 2,
-                          widget.currentQuestion.title,
-                          style: const TextStyle(
-                            color: Color.fromARGB(200, 0, 0, 0),
-                            fontSize: 20,
-                            fontWeight: FontWeight.w500,
-                            //height: 1.50,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  const SizedBox(
-                    height: 10,
-                  ),
-                  ListView.builder(
-                    shrinkWrap: true,
-                    itemBuilder: (context, index) {
-                      QuizQuestionOption option =
-                          widget.currentQuestion.options[index];
-                      return Padding(
-                        padding: const EdgeInsets.symmetric(
-                          vertical: 10,
-                          horizontal: 50,
-                        ),
-                        child: SizedBox(
-                          width: MediaQuery.of(context).size.width * 0.62,
-                          child: OptionsButton(
-                            buttonColor: const Color.fromARGB(69, 70, 168, 248),
-                            callBack: () {
-                              if (option.answer.isNotEmpty) {
-                                setState(() {
-                                  selectedOption = option;
-                                  widget.currentQuestion.options[index] =
-                                      option;
-                                  showAnswer = true;
-                                });
-                              }
-                            },
-                            text: option.title,
-                          ),
-                        ),
-                      );
-                    },
-                    itemCount: widget.currentQuestion.options.length,
-                  ),
-                  const SizedBox(
-                    height: 10,
-                  ),
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
 class QuizDraggingHandle extends StatelessWidget {
   const QuizDraggingHandle({super.key});
